fix ci test issue - import order
